Fix: Remove slow initial render time
* Removes slow initial render time by lazy loading avatars.
* Renders loading skeleton on top of avatar so that the image
  lazy loads in background (not displaying it will cause the
  image to never load).

Every so often, the render time for the
popup was very slow. This was always the
case for the first time you open.

To reproduce the error with 100% success,
I found that I needed to delete the browser cache.
Based on that, I knew it would be a networking issue.
I disabled avatars and found that the problem went away,
so the issue was image loading for avatars.
I confirmed that it was the sheer number of
avatars by slicing just 10, then 50, then 100
channel items, while observing that render
time increased depending on how many
avatars were being rendered.