Fix: Infinite loading state for Hoverable thumbnail
There was a race condition between setting
state for image loading flag and handling the
image onLoad event. To remove the race
condition, we don't render the image if the
popover is not open.

The tradeoff is that the image will load slightly
slower and will always have a short loading state,
but it is more consistent behavior.